apps/requirements.txt
aioredis==1.3.1
aliyun-python-sdk-core==2.13.19
aliyun-python-sdk-core-v3==2.13.11
aliyun-python-sdk-ecs==4.19.8
amqp==2.6.0
ansible==2.8.2
asgiref==3.2.10
asn1crypto==1.3.0
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.6.1
Automat==20.2.0
azure-common==1.1.25
azure-nspkg==3.0.2
azure-storage-blob==1.3.1
azure-storage-common==1.4.2
azure-storage-nspkg==3.1.0
backcall==0.2.0
bcrypt==3.1.7
billiard==3.6.3.0
boto==2.49.0
boto3==1.14.8
botocore==1.17.8
celery==4.4.5
certifi==2020.6.20
cffi==1.14.0
channels==2.4.0
channels-redis==2.4.2
chardet==3.0.4
configparser==5.0.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
crcmod==1.7
cryptography==2.9.2
daphne==2.5.0
decorator==4.4.2
deprecation==2.1.0
Django==2.2.10
django-auth-ldap==2.2.0
django-bootstrap3==14.0.0
django-cas-ng==4.1.1
django-celery-beat==2.0.0
django-filter==2.3.0
django-formtools==2.2
django-proxy==1.2.1
django-radius==1.5.0
django-ranged-response==0.2.0
django-redis==4.12.1
django-redis-cache==2.1.1
django-redis-sessions==0.6.1
django-rest-swagger==2.2.0
django-simple-captcha==0.5.12
django-timezone-field==4.0
djangorestframework==3.9.4
djangorestframework-bulk==0.2.1
dnspython==1.16.0
docutils==0.16
drf-nested-routers==0.91
drf-yasg==1.17.1
ecdsa==0.15
elasticsearch==6.4.0
enum-compat==0.0.3
ephem==3.7.7.1
eventlet==0.25.2
flower==0.9.4
ForgeryPy==0.1
future==0.18.2
greenlet==0.4.16
gunicorn==20.0.4
hiredis==1.0.1
httpsig==1.3.0
huaweicloud-sdk-python==1.0.22
humanize==0.5.1
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
inflection==0.5.0
ipip-ipdb==1.3.2
ipython==7.15.0
ipython-genutils==0.2.0
iso8601==0.1.12
itsdangerous==1.1.0
itypes==1.2.0
jedi==0.17.1
Jinja2==2.11.2
jmespath==0.10.0
jms-storage==0.0.29
jsonfield2==4.0.0.post0
jumpserver-django-oidc-rp==0.3.7.5
keystoneauth1==3.4.0
kombu==4.6.10
ldap3==2.7
lockfile==0.12.2
lxml==4.5.1
MarkupSafe==1.1.1
monotonic==1.5
msgpack==0.6.2
multi-key-dict==2.0.3
mysqlclient==1.4.6
netaddr==0.7.20
olefile==0.46
openapi-codec==1.3.2
oss2==2.4.0
packaging==20.4
paramiko==2.7.1
parso==0.7.0
passlib==1.7.2
pbr==5.4.5
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
prompt-toolkit==3.0.5
psutil==5.7.0
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycrypto==2.6.1
pycryptodome==3.9.7
pycryptodomex==3.9.7
Pygments==2.6.1
PyHamcrest==2.0.2
pyjwkest==1.4.2
PyNaCl==1.4.0
pyOpenSSL==19.1.0
pyotp==2.3.0
pyparsing==2.4.7
pyrad==2.3
python-cas==1.5.0
python-crontab==2.5.1
python-daemon==2.2.4
python-dateutil==2.8.1
python-gssapi==0.6.4
python-jenkins==1.7.0
python-ldap==3.3.0
python-redis-lock==3.5.0
pytz==2020.1
PyYAML==5.3.1
redis==3.5.3
requests==2.24.0
rest-condition==1.0.3
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
s3transfer==0.3.3
service-identity==18.1.0
simplejson==3.17.0
six==1.15.0
sqlparse==0.3.1
sshpubkeys==3.1.0
stevedore==2.0.1
tencentcloud-sdk-python==3.0.202
tornado==6.0.4
traitlets==4.3.3
treelib==1.5.3
Twisted==20.3.0
txaio==20.4.1
unicodecsv==0.14.1
uritemplate==3.0.1
urllib3==1.25.9
vine==1.3.0
wcwidth==0.2.4
Werkzeug==1.0.1
zope.interface==5.1.0
